remove manual lock for request counter, replace with executorservice's completedtaskcount + activecount

+++ src/main/java/com/example/demo/controller/MonitorController.java
@@ -21,7 +21,7 @@ public class MonitorController {
     }
 
     @GetMapping(value = "request_counter")
-    public int getRequestCounter() {
+    public long getRequestCounter() {
         return sleepAndSumService.getRequestCounter();
     }
 }
